requirements.txt
langchain
langchain-community 
langchain-core
streamlit
langchain_openai
beautifulsoup4
python_dotenv
faiss-cpu
pymupdf